src/ai_write_x/scrapers/pyproject.toml
[project]
name = "article-spider"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncpg>=0.29.0",
    "chardet>=5.2.0",
    "colorama==0.4.6",
    "cryptography>=45.0.7",
    "lxml>=6.0.1",
    "python-dotenv>=1.1.1",
]
